Fixing bugs

+++ image_backend/backend/views.py
@@ -88,14 +88,14 @@ def get_organisms_graphical_summary(request):
 
 
 def convert_to_radians(degrees):
-    return degrees * 3.14 / 180
+    return float(degrees) * 3.14 / 180
 
 
 def organisms_gis_search(request):
     filter_results = dict()
     filter_results['results'] = list()
-    latitude = convert_to_radians(float(request.GET.get('latitude', False)))
-    longitude = convert_to_radians(float(request.GET.get('longitude', False)))
+    latitude = convert_to_radians(request.GET.get('latitude', False))
+    longitude = convert_to_radians(request.GET.get('longitude', False))
     radius = float(request.GET.get('radius', False))
     results = SampleInfo.objects.filter(organisms__isnull=False)
     organisms = results.exclude(organisms__birth_location_longitude='',
@@ -171,9 +171,9 @@ def get_specimens_graphical_summary(request):
 def specimens_gis_search(request):
     filter_results = dict()
     filter_results['results'] = list()
-    latitude = convert_to_radians(float(request.GET.get('latitude', False)))
-    longitude = convert_to_radians(float(request.GET.get('longitude', False)))
-    radius = float(request.GET.get('radius', False))
+    latitude = convert_to_radians(request.GET.get('latitude', False))
+    longitude = convert_to_radians(request.GET.get('longitude', False))
+    radius = int(request.GET.get('radius', False))
     results = SampleInfo.objects.filter(specimens__isnull=False)
     specimens = results.exclude(specimens__collection_place_latitude='',
                                 specimens__collection_place_longitude='')
